Fix binary file reading bugs

+++ map_serializer.hpp
@@ -20,10 +20,10 @@
 using namespace std;
 
 namespace MapSerializer {
-    void map_serialize(const unordered_map<string, long long>&, string);
-    void map_deserialize(unordered_map<string, long long>&, string);
-    void triple_serialize(vector<tuple<long long, long long, long long> >&, string);
-    void triple_deserialize(vector<tuple<long long, long long, long long> >&, string);
+    void map_serialize(const unordered_map<string, unsigned int>&, string);
+    void map_deserialize(unordered_map<string, unsigned int>&, string);
+    void triple_serialize(vector<tuple<unsigned int, unsigned int, unsigned int> >&, string);
+    void triple_deserialize(vector<tuple<unsigned int, unsigned int, unsigned int> >&, string);
 }
 
 #endif /* map_serializer_hpp */
